Fixed broken registration in case of autocrop. Bumped version to 0.0.4.
This bug seems to be due to confusion about whether
imagePlus.duplicate() implicitly crops when the imagePlus object has a
ROI set on it. Interactively, a duplicate operation of an image with a
ROI set on it *does* crop. But simply using the ImagePlus API doesn't.
At least not on, for example, ImageJ 1.52n. Perhaps there was an
ImagePlus API change? Or was this a bug all along? Strange.